pyproject.toml
[project]
name = "grafana-sync"
version = "0.1.0"
description = "Sync resources between Grafana instances"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "argparse>=1.4.0",
    "logfmter>=0.0.8",
    "prometheus-client>=0.21.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
